requirements.txt
pyspark==3.5.1
findspark==2.0.1
sqlalchemy==2.0.29
psycopg2-binary==2.9.9
pandas==2.2.2
python-dotenv==1.0.1
